feat: support DD_SERVICE_MAPPING - resolves #1173
