+++ packages/dd-trace/src/util.js
@@ -10,7 +10,20 @@ function isFalse (str) {
   return str === 'false' || str === '0'
 }
 
+function toKeyValuePairs (str) {
+  return (str || '').split(',')
+    .filter(tag => tag.indexOf(':') !== -1)
+    .reduce((prev, next) => {
+      const tag = next.split(':')
+      const key = tag[0]
+      const value = tag.slice(1).join(':')
+      prev[key] = value
+      return prev
+    }, {})
+}
+
 module.exports = {
   isTrue,
-  isFalse
+  isFalse,
+  toKeyValuePairs
 }
